requirements/base.txt
Django==1.11.7
djangorestframework==3.7.3
mysqlclient==1.3.12
